requirements.txt
# Core dependencies
click>=8.1.0
pydantic>=2.0.0
rich>=13.0.0
pyyaml>=6.0.0
jinja2>=3.1.0
httpx>=0.25.0
python-dotenv>=1.0.0
deepdiff>=6.0.0
lxml>=5.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
ruff>=0.1.0
mypy>=1.6.0
pre-commit>=3.5.0
